requirements.txt
pyenchant
nltk
numpy
